pyproject.toml
[project]
name = "llm-gemini-code-assist"
version = "0.26.1"
description = "LLM plugin to access Google's Gemini via Code Assist API with OAuth"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Loki Coyote"},
    {name = "Simon Willison"}
]
license = "Apache-2.0"
classifiers = []
dependencies = [
    "llm>=0.27",
    "google-auth>=2.43.0",
    "requests>=2.32.5",
    "ijson>=3.4.0.post0",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "google-auth-oauthlib>=1.2.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest-dependency>=0.6.0",
    "pytest>=8.0.0,<9",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-recording>=0.13.4",
    "nest-asyncio>=1.6.0",
    "cogapp>=3.6.0",
    "google-auth-stubs>=0.3.0",
    "mypy>=1.14.0",
    "types-requests>=2.32.0",
    "ruff>=0.9.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/lokkju/llm-gemini-code-assist"
Issues = "https://github.com/lokkju/llm-gemini-code-assist/issues"

[project.entry-points.llm]
gemini_code_assist = "llm_gemini_code_assist"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
# Set line length to 100 (can adjust to 88 if you prefer black's default)
line-length = 100
target-version = "py310"

# Enable auto-fixing where possible
fix = true

[tool.ruff.lint]
# Select rule sets
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade (modern Python syntax)
    "B",    # flake8-bugbear (likely bugs)
    "A",    # flake8-builtins (shadowing builtins)
    "C4",   # flake8-comprehensions
    "PT",   # flake8-pytest-style
    "S",    # flake8-bandit (security)
    "RUF",  # ruff-specific rules
]

# Ignore specific rules if needed
ignore = [
    "S101",  # Use of assert (common in tests)
]

[tool.ruff.lint.per-file-ignores]
# Ignore security and assert checks in tests
"tests/**/*.py" = ["S", "S101"]

[tool.ruff.lint.isort]
# Configure import sorting
known-first-party = ["llm_gemini_code_assist", "tests"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
# Use double quotes by default
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
# Basic type checking configuration
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

# Per-module options
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false
